Funcionou essa meeeerda

+++ Sources/Salus.Communicator/src/main/java/bkd/src/salus/communicator/Domain/DTO/Medicine/MedicineDecrementRequest.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class MedicineDecrementRequest {
     private int UserId;
     private int MedicineId;
-    private int Quantity;
+    private float Quantity;
 }
